requirements.txt
# pip install -r requirements.txt

Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.0.1
Adafruit-WS2801==1.0.1
asn1crypto==0.24.0
configparser==3.5.0b2
cryptography==2.6.1
Django
dnspython==1.16.0
entrypoints==0.3
enum34==1.1.6
gpg==1.12.0
ipaddress==1.0.17
keyring==17.1.1
keyrings.alt==3.1.1
mysqlclient==1.3.10
pycrypto==2.6.1
PyGObject==3.30.4
pytz==2020.4
pyxdg==0.25
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
spidev==3.5
pygame